remove `FirstTracesInOrder`
