+++ dbux-data/src/applications/ApplicationSetData.js
@@ -1,128 +1,3 @@
-import EmptyArray from '@dbux/common/src/util/EmptyArray';
-import Trace from '@dbux/common/src/types/Trace';
-
-// ###########################################################################
-// FirstTracesInOrder
-// ###########################################################################
-
-class FirstTracesInOrder {
-  /**
-   * @type {Array<Trace>}
-   */
-  _firstTracesArray;
-  _firstTraceIndexById = new Map();
-
-  /**
-   * @param {ApplicationSetData} applicationSetData 
-   */
-  constructor(applicationSetData) {
-    this.applicationSetData = applicationSetData;
-    this.applicationSet = applicationSetData.set;
-    this._firstTracesArray = [];
-  }
-
-  _mergeAll() {
-    this._firstTracesArray = [];
-    const applications = this.applicationSetData.set.getAll();
-    const allFirstTraces = applications.map((app) => app.dataProvider.util.getFirstTracesInRuns() || EmptyArray);
-
-    // sort traces by trace.createdAt
-    const indexPointers = Array(applications.length).fill(0);
-    const tracesCount = allFirstTraces.reduce((sum, arr) => sum + arr.length, 0);
-
-    for (let i = 0; i < tracesCount; i++) {
-      let earliestTrace = null;
-      let earliestApplicationIndex = null;
-      for (let j = 0; j < applications.length; j++) {
-        const trace = allFirstTraces[j][indexPointers[j]];
-        if (!trace) continue;
-        if (!earliestTrace || trace.createdAt < earliestTrace.createdAt) {
-          earliestTrace = trace;
-          earliestApplicationIndex = j;
-        }
-      }
-      indexPointers[earliestApplicationIndex] += 1;
-      this._addOne(earliestTrace);
-    }
-  }
-
-  _handleApplicationsChanged = () => {
-    const applications = this.applicationSet.getAll();
-    this._mergeAll();
-
-    for (const app of applications) {
-      this.applicationSet.subscribe(
-        app.dataProvider.onData('executionContexts', this._addExecutionContexts.bind(this, app))
-      );
-    }
-  }
-
-  _addExecutionContexts(/* app, contexts */) {
-    // TODO: [performance] can we incrementally add new contexts only?
-    this._mergeAll();
-  }
-
-  _makeKey(firstTrace) {
-    const { applicationId } = firstTrace;
-    return `${applicationId}_${firstTrace.traceId}`;
-  }
-
-  _addOne = (firstTrace) => {
-    this._firstTracesArray.push(firstTrace);
-
-    const key = this._makeKey(firstTrace);
-    this._firstTraceIndexById.set(key, this._firstTracesArray.length - 1);
-  }
-
-  // ###########################################################################
-  // getters
-  // ###########################################################################
-
-  getAll() {
-    return this._firstTracesArray;
-  }
-
-  getIndex(firstTrace) {
-    const key = this._makeKey(firstTrace);
-    const index = this._firstTraceIndexById.get(key);
-    if (index === undefined) {
-      throw new Error('invalid query - given trace is not a root trace', firstTrace);
-    }
-    return index;
-  }
-
-  getFirstTraceInOrder() {
-    return this._firstTracesArray[0] || null;
-  }
-
-  /**
-   * @param {Trace} firstTrace 
-   */
-  getNextFirstTrace(firstTrace) {
-    const order = this.getIndex(firstTrace);
-    return this._firstTracesArray[order + 1] || null;
-  }
-
-  /**
-   * @param {Trace} firstTrace 
-   */
-  getPreviousFirstTrace(firstTrace) {
-    const order = this.getIndex(firstTrace);
-    return this._firstTracesArray[order - 1] || null;
-  }
-}
-
-
-// ###########################################################################
-// ApplicationSetData
-// ###########################################################################
-
-/**
- * @callback fileSelectedApplicationCallback
- * @param {Application} application
- * @param {number} programId
- */
-
 /**
  * Encapsulates all data that is related to the set of selected applications;
  * specifically, any data that changes when selected applications change.
@@ -132,7 +7,10 @@ class FirstTracesInOrder {
 export default class ApplicationSetData {
   constructor(applicationSet) {
     this.applicationSet = applicationSet;
-    this.firstTracesInOrder = new FirstTracesInOrder(this);
+    // // TODO
+    // this.threadSelection; // { _all: [threadId] }
+    // this.rootContextsInOrder; // listen on selection
+    // this.firstTracesInOrder = new FirstTracesInOrder(this);
 
     // this.applicationSet._emitter.on('_applicationsChanged0', this._handleApplicationsChanged);
     this.applicationSet.onApplicationsChanged(this._handleApplicationsChanged);
@@ -143,7 +21,7 @@ export default class ApplicationSetData {
   }
 
   _handleApplicationsChanged = () => {
-    this.firstTracesInOrder._handleApplicationsChanged();
+    // this.firstTracesInOrder._handleApplicationsChanged();
   }
 
   getTrace(applicationId, traceId) {
@@ -161,6 +39,12 @@ export default class ApplicationSetData {
       return sum + !!programId;
     }, 0);
   }
+
+  /**
+   * @callback fileSelectedApplicationCallback
+   * @param {Application} application
+   * @param {number} programId
+   */
 
   /**
    * @param {fileSelectedApplicationCallback} cb
